+++ providers/go-feature-flag-in-process/go.mod
@@ -1,0 +1,24 @@
+module github.com/open-feature/go-sdk-contrib/providers/go-feature-flag-in-process
+
+go 1.22.5
+
+require (
+	github.com/open-feature/go-sdk v1.12.0
+	github.com/thomaspoignant/go-feature-flag v1.31.2
+)
+
+require (
+	github.com/BurntSushi/toml v1.4.0 // indirect
+	github.com/antlr4-go/antlr/v4 v4.13.0 // indirect
+	github.com/blang/semver v3.5.1+incompatible // indirect
+	github.com/bluele/gcache v0.0.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-logr/logr v1.4.2 // indirect
+	github.com/google/go-cmp v0.6.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/nikunjy/rules v1.5.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.12.0 // indirect
+	golang.org/x/exp v0.0.0-20240506185415-9bf2ced13842 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
