+++ providers/go-feature-flag/pkg/controller/data_collector_manager.go
@@ -2,7 +2,7 @@ package controller
 
 import (
 	"fmt"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/provider_v2/model"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/model"
 	"sync"
 	"time"
 )
