+++ providers/go-feature-flag/pkg/controller/data_collector_manager_test.go
@@ -1,8 +1,8 @@
 package controller_test
 
 import (
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/provider_v2/controller"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/provider_v2/model"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/controller"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/model"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"testing"
