+++ providers/go-feature-flag/pkg/hook/data_collector_hook.go
@@ -2,8 +2,8 @@ package hook
 
 import (
 	"context"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/provider_v2/controller"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/provider_v2/model"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/controller"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/model"
 	"github.com/open-feature/go-sdk/openfeature"
 	"time"
 )
