--- providers/go-feature-flag/pkg/model/eval_response.go
@@ -1,12 +1,0 @@
-package model
-
-type EvalResponse[T JsonType] struct {
-	TrackEvents   bool   `json:"trackEvents"`
-	VariationType string `json:"variationType"`
-	Failed        bool   `json:"failed"`
-	Version       string `json:"version"`
-	Reason        string `json:"reason"`
-	ErrorCode     string `json:"errorCode"`
-	Value         T      `json:"value"`
-	Cacheable     bool   `json:"cacheable"`
-}
