--- providers/go-feature-flag/pkg/model/generic_evaluation_detail.go
@@ -1,8 +1,0 @@
-package model
-
-import of "github.com/open-feature/go-sdk/openfeature"
-
-type GenericResolutionDetail[T JsonType] struct {
-	Value T
-	of.ProviderResolutionDetail
-}
