--- providers/go-feature-flag/pkg/model/json_type.go
@@ -1,5 +1,0 @@
-package model
-
-type JsonType interface {
-	float64 | int64 | string | bool | interface{}
-}
