+++ providers/go-feature-flag/pkg/provider.go
@@ -1,58 +1,32 @@
 package gofeatureflag
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"github.com/bluele/gcache"
-	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/model"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/controller"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/hook"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/util"
+	"github.com/open-feature/go-sdk-contrib/providers/ofrep"
 	of "github.com/open-feature/go-sdk/openfeature"
-	client "github.com/thomaspoignant/go-feature-flag"
-	"github.com/thomaspoignant/go-feature-flag/exporter"
-	"github.com/thomaspoignant/go-feature-flag/exporter/webhookexporter"
-	"github.com/thomaspoignant/go-feature-flag/ffcontext"
-	"io"
-	"net/http"
-	"net/url"
-	"path"
 	"time"
 )
 
-const defaultCacheSize = 10000
-const defaultCacheTTL = 1 * time.Minute
-const defaultDataCacheMaxEventInMemory = 500
-const defaultDataCacheFlushInterval = 1 * time.Minute
+const providerName = "GO Feature Flag"
+const cacheableMetadataKey = "gofeatureflag_cacheable"
 
-// Provider is the OpenFeature provider for GO Feature Flag.
 type Provider struct {
-	httpClient             HTTPClient
-	endpoint               string
-	goFeatureFlagInstance  *client.GoFeatureFlag
-	apiKey                 string
-	cache                  gcache.Cache
-	cacheTTL               time.Duration
-	cacheDisable           bool
-	dataCollectorScheduler *exporter.Scheduler
-}
-
-// HTTPClient is a custom interface to be able to override it by any implementation
-// of an HTTP client.
-type HTTPClient interface {
-	Do(req *http.Request) (*http.Response, error)
-}
-
-// defaultHTTPClient is the default HTTP client used to call GO Feature Flag.
-// By default, we have a timeout of 10000 milliseconds.
-func defaultHTTPClient() HTTPClient {
-	netTransport := &http.Transport{
-		TLSHandshakeTimeout: 10000 * time.Millisecond,
+	ofrepProvider        *ofrep.Provider
+	cache                *controller.Cache
+	dataCollectorManager controller.DataCollectorManager
+	options              ProviderOptions
+	status               of.State
+	hooks                []of.Hook
+	goffAPI              controller.GoFeatureFlagAPI
+	pollingInfo          struct {
+		ticker  *time.Ticker
+		channel chan bool
 	}
-
-	return &http.Client{
-		Timeout:   10000 * time.Millisecond,
-		Transport: netTransport,
-	}
+	events chan of.Event
 }
 
 // NewProvider allows you to create a GO Feature Flag provider without any context.
@@ -63,426 +37,225 @@ func NewProvider(options ProviderOptions) (*Provider, error) {
 
 // NewProviderWithContext is the easiest way of creating a new GO Feature Flag provider.
 func NewProviderWithContext(ctx context.Context, options ProviderOptions) (*Provider, error) {
-	if options.GOFeatureFlagConfig != nil {
-		goff, err := client.New(*options.GOFeatureFlagConfig)
-		if err != nil {
-			return nil, err
-		}
-		return &Provider{
-			goFeatureFlagInstance: goff,
-		}, nil
+	if err := options.Validation(); err != nil {
+		return nil, err
 	}
-
-	if options.Endpoint == "" {
-		return nil, fmt.Errorf("invalid provider options, empty endpoint value")
+	ofrepOptions := make([]ofrep.Option, 0)
+	if options.APIKey != "" {
+		ofrepOptions = append(ofrepOptions, ofrep.WithBearerToken(options.APIKey))
 	}
-
-	// Set default values
-	httpClient := options.HTTPClient
-	if httpClient == nil {
-		httpClient = defaultHTTPClient()
+	if options.HTTPClient != nil {
+		ofrepOptions = append(ofrepOptions, ofrep.WithClient(options.HTTPClient))
 	}
-	if options.FlagCacheSize == 0 {
-		options.FlagCacheSize = defaultCacheSize
-	}
-	if options.DataMaxEventInMemory == 0 {
-		options.DataMaxEventInMemory = defaultDataCacheMaxEventInMemory
-	}
-	if options.DataFlushInterval == 0 {
-		options.DataFlushInterval = defaultDataCacheFlushInterval
-	}
-	if options.FlagCacheTTL == 0 {
-		options.FlagCacheTTL = defaultCacheTTL
-	}
-	scheduler := startDataCollector(ctx, options)
+	ofrepProvider := ofrep.NewProvider(options.Endpoint, ofrepOptions...)
+	cacheCtrl := controller.NewCache(options.FlagCacheSize, options.FlagCacheTTL, options.DisableCache)
+	goffAPI := controller.NewGoFeatureFlagAPI(controller.GoFeatureFlagApiOptions{
+		Endpoint:   options.Endpoint,
+		HTTPClient: options.HTTPClient,
+		APIKey:     options.APIKey,
+	})
+	dataCollectorManager := controller.NewDataCollectorManager(
+		goffAPI,
+		options.DataCollectorMaxEventStored,
+		options.DataFlushInterval,
+	)
 	return &Provider{
-		apiKey:                 options.APIKey,
-		endpoint:               options.Endpoint,
-		httpClient:             httpClient,
-		cacheTTL:               options.FlagCacheTTL,
-		cacheDisable:           options.DisableCache,
-		cache:                  gcache.New(options.FlagCacheSize).LRU().Build(),
-		dataCollectorScheduler: scheduler,
+		ofrepProvider:        ofrepProvider,
+		cache:                cacheCtrl,
+		dataCollectorManager: dataCollectorManager,
+		options:              options,
+		goffAPI:              goffAPI,
+		events:               make(chan of.Event, 5),
 	}, nil
 }
-func startDataCollector(ctx context.Context, options ProviderOptions) *exporter.Scheduler {
-	if options.DisableCache {
-		return nil
-	}
 
-	u, _ := url.Parse(options.Endpoint)
-	u.Path = path.Join(u.Path, "v1", "/")
-	u.Path = path.Join(u.Path, "data", "/")
-	u.Path = path.Join(u.Path, "collector", "/")
-
-	webhook := webhookexporter.Exporter{
-		EndpointURL: u.String(),
-		Meta: map[string]string{
-			"provider":    "go",
-			"openfeature": "true",
-		},
-	}
-	if options.APIKey != "" {
-		webhook.Headers = map[string][]string{
-			"Authorization": {fmt.Sprintf("Bearer %s", options.APIKey)},
-		}
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	scheduler := exporter.NewScheduler(ctx,
-		options.DataFlushInterval, options.DataMaxEventInMemory, &webhook, nil)
-	go scheduler.StartDaemon()
-	return scheduler
-}
-
-// Metadata returns the meta of the GO Feature Flag provider.
 func (p *Provider) Metadata() of.Metadata {
 	return of.Metadata{
-		Name: "GO Feature Flag",
+		Name: fmt.Sprintf("%s Provider", providerName),
 	}
 }
 
 func (p *Provider) BooleanEvaluation(ctx context.Context, flag string, defaultValue bool, evalCtx of.FlattenedContext) of.BoolResolutionDetail {
-	res := genericEvaluation[bool](p, ctx, flag, defaultValue, evalCtx)
-	return of.BoolResolutionDetail{
-		Value:                    res.Value,
-		ProviderResolutionDetail: res.ProviderResolutionDetail,
+	if err := util.ValidateTargetingKey(evalCtx); err != nil {
+		return of.BoolResolutionDetail{
+			Value:                    defaultValue,
+			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
+		}
 	}
+	if cacheValue, err := p.cache.GetBool(flag, evalCtx); err == nil && cacheValue != nil {
+		cacheValue.Reason = of.CachedReason
+		return *cacheValue
+	}
+	res := p.ofrepProvider.BooleanEvaluation(ctx, flag, defaultValue, evalCtx)
+	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
+		_ = p.cache.Set(flag, evalCtx, res)
+	}
+	return res
 }
+
 func (p *Provider) StringEvaluation(ctx context.Context, flag string, defaultValue string, evalCtx of.FlattenedContext) of.StringResolutionDetail {
-	res := genericEvaluation[string](p, ctx, flag, defaultValue, evalCtx)
-	return of.StringResolutionDetail{
-		Value:                    res.Value,
-		ProviderResolutionDetail: res.ProviderResolutionDetail,
+	if err := util.ValidateTargetingKey(evalCtx); err != nil {
+		return of.StringResolutionDetail{
+			Value:                    defaultValue,
+			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
+		}
 	}
+	if cacheValue, err := p.cache.GetString(flag, evalCtx); err == nil && cacheValue != nil {
+		cacheValue.Reason = of.CachedReason
+		return *cacheValue
+	}
+	res := p.ofrepProvider.StringEvaluation(ctx, flag, defaultValue, evalCtx)
+	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
+		_ = p.cache.Set(flag, evalCtx, res)
+	}
+	return res
 }
+
 func (p *Provider) FloatEvaluation(ctx context.Context, flag string, defaultValue float64, evalCtx of.FlattenedContext) of.FloatResolutionDetail {
-	res := genericEvaluation[float64](p, ctx, flag, defaultValue, evalCtx)
-	return of.FloatResolutionDetail{
-		Value:                    res.Value,
-		ProviderResolutionDetail: res.ProviderResolutionDetail,
+	if err := util.ValidateTargetingKey(evalCtx); err != nil {
+		return of.FloatResolutionDetail{
+			Value:                    defaultValue,
+			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
+		}
 	}
+	if cacheValue, err := p.cache.GetFloat(flag, evalCtx); err == nil && cacheValue != nil {
+		cacheValue.Reason = of.CachedReason
+		return *cacheValue
+	}
+	res := p.ofrepProvider.FloatEvaluation(ctx, flag, defaultValue, evalCtx)
+	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
+		_ = p.cache.Set(flag, evalCtx, res)
+	}
+	return res
 }
+
 func (p *Provider) IntEvaluation(ctx context.Context, flag string, defaultValue int64, evalCtx of.FlattenedContext) of.IntResolutionDetail {
-	res := genericEvaluation[int64](p, ctx, flag, defaultValue, evalCtx)
-	return of.IntResolutionDetail{
-		Value:                    res.Value,
-		ProviderResolutionDetail: res.ProviderResolutionDetail,
+	if err := util.ValidateTargetingKey(evalCtx); err != nil {
+		return of.IntResolutionDetail{
+			Value:                    defaultValue,
+			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
+		}
 	}
+	if cacheValue, err := p.cache.GetInt(flag, evalCtx); err == nil && cacheValue != nil {
+		cacheValue.Reason = of.CachedReason
+		return *cacheValue
+	}
+	res := p.ofrepProvider.IntEvaluation(ctx, flag, defaultValue, evalCtx)
+	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
+		_ = p.cache.Set(flag, evalCtx, res)
+	}
+	return res
 }
+
 func (p *Provider) ObjectEvaluation(ctx context.Context, flag string, defaultValue interface{}, evalCtx of.FlattenedContext) of.InterfaceResolutionDetail {
-	res := genericEvaluation[interface{}](p, ctx, flag, defaultValue, evalCtx)
-	return of.InterfaceResolutionDetail{
-		Value:                    res.Value,
-		ProviderResolutionDetail: res.ProviderResolutionDetail,
+	if err := util.ValidateTargetingKey(evalCtx); err != nil {
+		return of.InterfaceResolutionDetail{
+			Value:                    defaultValue,
+			ProviderResolutionDetail: of.ProviderResolutionDetail{ResolutionError: *err, Reason: of.ErrorReason},
+		}
 	}
+	if cacheValue, err := p.cache.GetInterface(flag, evalCtx); err == nil && cacheValue != nil {
+		cacheValue.Reason = of.CachedReason
+		return *cacheValue
+	}
+	res := p.ofrepProvider.ObjectEvaluation(ctx, flag, defaultValue, evalCtx)
+	if cachable, err := res.FlagMetadata.GetBool(cacheableMetadataKey); err == nil && cachable {
+		_ = p.cache.Set(flag, evalCtx, res)
+	}
+	return res
 }
 
-func (p *Provider) Shutdown() {
-	p.dataCollectorScheduler.Close()
-}
-
-// Hooks is returning an empty array because GO Feature Flag does not use any hooks.
 func (p *Provider) Hooks() []of.Hook {
-	return []of.Hook{}
+	return p.hooks
 }
 
-// genericEvaluation is doing evaluation for all types using generics.
-func genericEvaluation[T model.JsonType](provider *Provider, ctx context.Context, flagName string, defaultValue T, evalCtx of.FlattenedContext) model.GenericResolutionDetail[T] {
-	goffRequestBody, errConvert := model.NewEvalFlagRequest[T](evalCtx, defaultValue)
-	if errConvert != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: *errConvert,
-				Reason:          of.ErrorReason,
-			},
-		}
+// Init holds initialization logic of the provider
+func (p *Provider) Init(_ of.EvaluationContext) error {
+	if !p.options.DisableDataCollector {
+		dataCollectorHook := hook.NewDataCollectorHook(&p.dataCollectorManager)
+		p.hooks = []of.Hook{dataCollectorHook}
+		p.dataCollectorManager.Start()
 	}
 
-	// if we have a GO Feature Flag instance instantiate we evaluate the flag locally,
-	// using the GO module directly. We will not send any remote calls to the relay proxy.
-	if provider.goFeatureFlagInstance != nil {
-		return evaluateLocally(provider, goffRequestBody, flagName, defaultValue)
+	// Start polling to check if there is any flag change in order to invalidate the cache.
+	if p.options.FlagChangePollingInterval >= 0 && !p.options.DisableCache {
+		p.startPolling(p.options.FlagChangePollingInterval)
 	}
-	return evaluateWithRelayProxy(provider, ctx, goffRequestBody, flagName, defaultValue)
+
+	p.status = of.ReadyState
+	p.events <- of.Event{
+		ProviderName: providerName, EventType: of.ProviderReady,
+		ProviderEventDetails: of.ProviderEventDetails{Message: "Provider is ready"}}
+	return nil
 }
 
-// evaluateLocally is using the GO Feature Flag module to evaluate your flag.
-// it means that you don't need any relay proxy to make it work.
-func evaluateLocally[T model.JsonType](provider *Provider, goffRequestBody model.EvalFlagRequest, flagName string, defaultValue T) model.GenericResolutionDetail[T] {
-	// Construct user
-	ctxBuilder := ffcontext.NewEvaluationContextBuilder(goffRequestBody.EvaluationContext.Key)
-	for k, v := range goffRequestBody.EvaluationContext.Custom {
-		ctxBuilder.AddCustom(k, v)
-	}
-
-	// Call GO Module
-	rawResult, err := provider.goFeatureFlagInstance.RawVariation(flagName, ctxBuilder.Build(), defaultValue)
-	if err != nil {
-		switch rawResult.ErrorCode {
-		case string(of.FlagNotFoundCode):
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewFlagNotFoundResolutionError(fmt.Sprintf("flag %s was not found in GO Feature Flag", flagName)),
-					Reason:          of.ErrorReason,
-				},
-			}
-		case string(of.ProviderNotReadyCode):
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewProviderNotReadyResolutionError(
-						fmt.Sprintf("provider not ready for evaluation of flag %s", flagName)),
-					Reason: of.ErrorReason,
-				},
-			}
-		case string(of.ParseErrorCode):
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewParseErrorResolutionError(
-						fmt.Sprintf("parse error during evaluation of flag %s", flagName)),
-					Reason: of.ErrorReason,
-				},
-			}
-		case string(of.TypeMismatchCode):
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewTypeMismatchResolutionError(
-						fmt.Sprintf("unexpected type for flag %s", flagName)),
-					Reason: of.ErrorReason,
-				},
-			}
-		case string(of.GeneralCode):
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewGeneralResolutionError(
-						fmt.Sprintf("unexpected error during evaluation of the flag %s", flagName)),
-					Reason: of.ErrorReason,
-				},
-			}
-		}
-	}
-
-	// This part convert the int received by the module to int64 to be compatible with
-	// the types expect by Open-feature.
-	var v model.JsonType
-	switch value := rawResult.Value.(type) {
-	case int:
-		v = int64(value)
-	default:
-		v = value
-	}
-
-	switch value := v.(type) {
-	case nil:
-		return model.GenericResolutionDetail[T]{
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				Reason:  of.Reason(rawResult.Reason),
-				Variant: rawResult.VariationType,
-			},
-		}
-	case T:
-		return model.GenericResolutionDetail[T]{
-			Value: value,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				Reason:  of.Reason(rawResult.Reason),
-				Variant: rawResult.VariationType,
-			},
-		}
-	default:
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewTypeMismatchResolutionError(fmt.Sprintf("unexpected type for flag %s", flagName)),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
+// Status exposes the status of the provider
+func (p *Provider) Status() of.State {
+	return p.status
 }
 
-func convertCache[T model.JsonType](value interface{}) (model.GenericResolutionDetail[T], error) {
-	switch v := value.(type) {
-	case model.GenericResolutionDetail[T]:
-		return v, nil
-	default:
-		return model.GenericResolutionDetail[T]{}, fmt.Errorf("impossible to convert into the cache")
+// Shutdown defines the shutdown operation of the provider
+func (p *Provider) Shutdown() {
+	if !p.options.DisableDataCollector {
+		p.hooks = []of.Hook{}
+		p.dataCollectorManager.Stop()
 	}
+	p.stopPolling()
 }
 
-// evaluateWithRelayProxy is calling GO Feature Flag relay proxy to evaluate the file.
-func evaluateWithRelayProxy[T model.JsonType](provider *Provider, ctx context.Context, goffRequestBody model.EvalFlagRequest, flagName string, defaultValue T) model.GenericResolutionDetail[T] {
-	cacheKey := fmt.Sprintf("%s-%+v", flagName, goffRequestBody.EvaluationContext)
-	// check if flag is available in the cache
-	cacheResInterface, err := provider.cache.Get(cacheKey)
-	if err == nil {
-		// we have retrieve something from the cache.
-		cacheValue, err := convertCache[T](cacheResInterface)
-		if err != nil {
-			// impossible to convert the cache, we remove the entry from the cache assuming the next
-			// call to convertCache wouldn't result in the same error on the next call.
-			provider.cache.Remove(cacheKey)
-		} else {
-			event := exporter.NewFeatureEvent(
-				ffcontext.NewEvaluationContext(goffRequestBody.EvaluationContext.Key),
-				flagName,
-				cacheValue.Value,
-				cacheValue.Variant,
-				cacheValue.Reason == of.ErrorReason,
-				"",
-				"PROVIDER_CACHE",
-			)
-			provider.dataCollectorScheduler.AddEvent(event)
-			cacheValue.Reason = of.CachedReason
-			return cacheValue
-		}
-	}
+// EventChannel returns the event channel of this provider
+func (p *Provider) EventChannel() <-chan of.Event {
+	return p.events
+}
 
-	goffRequestBodyStr, err := json.Marshal(goffRequestBody)
-	if err != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError("impossible to marshal GO Feature Flag request"),
-				Reason:          of.ErrorReason,
-			},
-		}
+// startPolling starts the polling mechanism that checks if the configuration has changed.
+func (p *Provider) startPolling(pollingInterval time.Duration) {
+	if pollingInterval == 0 {
+		pollingInterval = 120000 * time.Millisecond
 	}
+	p.pollingInfo = struct {
+		ticker  *time.Ticker
+		channel chan bool
+	}{
+		ticker:  time.NewTicker(pollingInterval),
+		channel: make(chan bool),
+	}
+	go func() {
+		for {
+			select {
+			case <-p.pollingInfo.channel:
+				return
+			case <-p.pollingInfo.ticker.C:
+				changeStatus, err := p.goffAPI.ConfigurationHasChanged()
+				switch changeStatus {
+				case controller.FlagConfigurationInitialized,
+					controller.FlagConfigurationNotChanged:
+					// do nothing
 
-	evalURL, err := url.Parse(provider.endpoint)
-	if err != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError("impossible to parse GO Feature Flag endpoint option"),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
-	evalURL.Path = path.Join(evalURL.Path, "v1", "/")
-	evalURL.Path = path.Join(evalURL.Path, "feature", "/")
-	evalURL.Path = path.Join(evalURL.Path, flagName, "/")
-	evalURL.Path = path.Join(evalURL.Path, "eval", "/")
-
-	goffRequest, err :=
-		http.NewRequestWithContext(ctx, http.MethodPost, evalURL.String(), bytes.NewBuffer(goffRequestBodyStr))
-	if err != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError("error while building GO Feature Flag relay proxy request"),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
-	goffRequest.Header.Set("Content-Type", "application/json")
-	if provider.apiKey != "" {
-		goffRequest.Header.Set("Authorization", fmt.Sprintf("Bearer %s", provider.apiKey))
-	}
-
-	response, err := provider.httpClient.Do(goffRequest)
-	if err != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError("impossible to contact GO Feature Flag relay proxy instance"),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
-	responseStr, err := io.ReadAll(response.Body)
-	if err != nil {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError("impossible to read API response from GO Feature Flag"),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
-
-	if response.StatusCode == http.StatusUnauthorized {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError(
-					"invalid token used to contact GO Feature Flag relay proxy instance"),
-				Reason: of.ErrorReason,
-			},
-		}
-	}
-	if response.StatusCode >= http.StatusBadRequest {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewGeneralResolutionError(
-					"unexpected answer from the relay proxy"),
-				Reason: of.ErrorReason,
-			},
-		}
-	}
-
-	var evalResponse model.EvalResponse[T]
-	err = json.Unmarshal(responseStr, &evalResponse)
-	if err != nil {
-		if err.Error() == "unexpected end of JSON input" {
-			return model.GenericResolutionDetail[T]{
-				Value: defaultValue,
-				ProviderResolutionDetail: of.ProviderResolutionDetail{
-					ResolutionError: of.NewParseErrorResolutionError(
-						fmt.Sprintf("impossible to parse response for flag %s: %s", flagName, responseStr)),
-					Reason: of.ErrorReason,
-				},
+				case controller.FlagConfigurationUpdated:
+					// Clearing the cache when the configuration is updated
+					p.cache.Purge()
+					p.events <- of.Event{
+						ProviderName: providerName, EventType: of.ProviderConfigChange,
+						ProviderEventDetails: of.ProviderEventDetails{Message: "Configuration has changed"}}
+				case controller.ErrorConfigurationChange:
+					p.events <- of.Event{
+						ProviderName: providerName, EventType: of.ProviderStale,
+						ProviderEventDetails: of.ProviderEventDetails{
+							Message: "Impossible to check configuration change " + err.Error()},
+					}
+				}
 			}
 		}
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewTypeMismatchResolutionError(fmt.Sprintf("unexpected type for flag %s", flagName)),
-				Reason:          of.ErrorReason,
-			},
-		}
-	}
+	}()
+}
 
-	if evalResponse.ErrorCode == string(of.FlagNotFoundCode) {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				ResolutionError: of.NewFlagNotFoundResolutionError(fmt.Sprintf("flag %s was not found in GO Feature Flag", flagName)),
-				Reason:          of.ErrorReason,
-			},
-		}
+// stopPolling stops the polling mechanism that check if the configuration has changed.
+func (p *Provider) stopPolling() {
+	if p.pollingInfo.channel != nil {
+		p.pollingInfo.channel <- true
 	}
-
-	if evalResponse.Reason == string(of.DisabledReason) {
-		return model.GenericResolutionDetail[T]{
-			Value: defaultValue,
-			ProviderResolutionDetail: of.ProviderResolutionDetail{
-				Reason:  of.DisabledReason,
-				Variant: "SdkDefault",
-			},
-		}
+	if p.pollingInfo.ticker != nil {
+		p.pollingInfo.ticker.Stop()
 	}
-
-	resDetail := model.GenericResolutionDetail[T]{
-		Value: evalResponse.Value,
-		ProviderResolutionDetail: of.ProviderResolutionDetail{
-			Reason:  of.Reason(evalResponse.Reason),
-			Variant: evalResponse.VariationType,
-		},
-	}
-
-	if !provider.cacheDisable && evalResponse.Cacheable {
-		if provider.cacheTTL == -1 {
-			_ = provider.cache.Set(cacheKey, resDetail)
-		} else {
-			_ = provider.cache.SetWithExpire(cacheKey, resDetail, provider.cacheTTL)
-		}
-	}
-	return resDetail
 }
