+++ providers/go-feature-flag/pkg/provider_options.go
@@ -1,7 +1,10 @@
 package gofeatureflag
 
 import (
+	"fmt"
+	"github.com/open-feature/go-sdk-contrib/providers/go-feature-flag/pkg/goff_error"
 	ffclient "github.com/thomaspoignant/go-feature-flag"
+	"net/http"
 	"time"
 )
 
@@ -14,7 +17,7 @@ type ProviderOptions struct {
 
 	// HTTPClient (optional) is the HTTP Client we will use to contact GO Feature Flag.
 	// By default, we are using a custom HTTPClient with a timeout configure to 10000 milliseconds.
-	HTTPClient HTTPClient
+	HTTPClient *http.Client
 
 	// GOFeatureFlagConfig is the configuration struct for the GO Feature Flag module.
 	// If not nil we will launch the provider using the GO Feature Flag module.
@@ -50,4 +53,25 @@ type ProviderOptions struct {
 	// when calling the evaluation API.
 	// default: 500
 	DataMaxEventInMemory int64
+
+	// DataCollectorMaxEventStored (optional) maximum number of event we keep in memory, if we reach this number it means
+	// that we will start to drop the new events. This is a security to avoid a memory leak.
+	// default: 100000
+	DataCollectorMaxEventStored int64
+
+	// DisableDataCollector (optional) set to true if you would like to disable the data collector.
+	DisableDataCollector bool
+
+	// FlagChangePollingInterval (optional) interval time we poll the proxy to check if the configuration has changed.
+	// If the cache is enabled, we will poll the relay-proxy every X milliseconds to check if the configuration has changed.
+	// Use -1 if you want to deactivate polling.
+	// default: 120000ms
+	FlagChangePollingInterval time.Duration
+}
+
+func (o *ProviderOptions) Validation() error {
+	if o.Endpoint == "" {
+		return goff_error.NewInvalidOption(fmt.Sprintf("invalid option: %s", o.Endpoint))
+	}
+	return nil
 }
